Now can submit zip files, with or without makefiles. Upper limit for zip file and extracted files have been set.

+++ run_code.js
@@ -16,34 +16,39 @@ function stdPath(filePath){
 }
 
 // spawns a process and returns the process object
-async function execCode(filePath, lang){
+async function execCode(filePath, lang, mode){
 	const dirname = path.dirname(filePath);
 	const filename = path.basename(filePath);
 	try{
-		if (lang === "python"){
+		if (Number(mode) === 0){
+				if (lang === "python"){
+					const {stdout, stderr} = await execPromise(`docker run --rm `+
+															   `-v ${dirname}:/usr/src/app -w /usr/src/app `+
+															   `--memory="128m" --cpus="1.0" `+
+																`--ulimit fsize=${25 * 1024 * 1024} --ulimit core=0 --pids-limit 20 `+
+															   `python:3.10-slim python3 ${filename}`);
+					return {stdout : stdout, stderr : stderr};
+				}
+				else if (lang === "C" || lang === "C++"){
+					const compiler = lang === "C" ? "gcc" : "g++";
+					const outPath = path.join(dirname, 'a.out');
+					//compile and execute
+					const {stdout, stderr} = await execPromise(
+											`docker run --rm -v ${dirname}:/usr/src/app -w /usr/src/app `+
+											`--memory="128m" --cpus="1.0" `+
+											`--ulimit fsize=${25 * 1024 * 1024} --ulimit core=0 --pids-limit 20 `+
+											`gcc:latest bash -c "${compiler} -Wall ${filename} -o a.out && ./a.out"`
+											);
+					return {stdout : stdout, stderr : stderr};
+				}
+		}else if (Number(mode) === 1){
 			const {stdout, stderr} = await execPromise(`docker run --rm `+
-													   `-v ${dirname}:/usr/src/app -w /usr/src/app `+
-													   `--memory="128m" --cpus="1.0" `+
-													   `python:3.10-slim python3 ${filename}`);
+									`-v ${dirname}:/usr/src/app -w /usr/src/app `+
+									`--memory="128m" --cpus="1.0"`+
+									`--ulimit fsize=${25 * 1024 * 1024} --ulimit core=0 --pids-limit 20 `+
+									`gcc:latest make`);	
 			return {stdout : stdout, stderr : stderr};
-		}
-		else if (lang === "C" || lang === "C++"){
-			const compiler = lang === "C" ? "gcc" : "g++";
-			const outPath = path.join(dirname, 'a.out');
-			//compile and execute
-			const {stdout, stderr} = await execPromise(
-									`docker run --rm -v ${dirname}:/usr/src/app -w /usr/src/app `+
-									`--memory="128m" --cpus="1.0" `+
-									`gcc:latest bash -c ${compiler} -Wall ${filename} -o a.out && ./a.out`
-									);
-			return {stdout : stdout, stderr : stderr};
-		}
-		else{
-			// remove directory
-			fs.rm(dirname, {recursive : true},(err)=>{
-				if (err) 
-					throw new Error(err.message);
-			});
+		}else{
 			throw new Error(`Execution for language ${lang} is not configured.`);
 		}
 	}catch(err){
